Deleted studentController file to prevent confusion.

+++ src/main/java/io/ims/backend/Models/Student.java
@@ -27,7 +27,7 @@ public class Student extends UserDetails{
    private Long courseID;
    private String section;
 
-    public Student(String email, String password, String userRole, String firstName, String lastName, String gender, Integer age, String birthDate, String homeAddress, Integer contactNumber, String civilStatus, String yearLevel, Integer courseID, String section) {
+    public Student(String email, String password, String userRole, String firstName, String lastName, String gender, Integer age, String birthDate, String homeAddress, Integer contactNumber, String civilStatus, String yearLevel, Long courseID, String section) {
         super(email, password, userRole, firstName, lastName, gender, age, birthDate, homeAddress, contactNumber, civilStatus);
         this.yearLevel = yearLevel;
         this.courseID = courseID;
